services/auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
python-jose
email-validator
pydantic-settings
PyJWT
